emb-1300: use plugin for handleLink in eajs and fix double sdk plugin init
