+++ enterprise/frontend/src/metabase-enterprise/sdk-iframe-embedding-plugins.ts
@@ -2,6 +2,11 @@ import { initializePlugins } from "sdk-ee-plugins";
 
 // Plugins that are only applicable to the new iframe embedding.
 import { initializePlugin as initializeEmbeddingIframeSdkPlugin } from "./embedding_iframe_sdk";
+import { initializeHandleLinkPlugin } from "./embedding_iframe_sdk/handle-link";
 
-initializeEmbeddingIframeSdkPlugin();
+// Sdk plugins plugins
 initializePlugins();
+
+// EAJS specific plugins and overrides
+initializeEmbeddingIframeSdkPlugin();
+initializeHandleLinkPlugin();
