+++ frontend/test/metabase/lib/dom.unit.spec.js
@@ -1,3 +1,5 @@
+import { setupSdkPlugins } from "__support__/enterprise";
+import { mockSettings } from "__support__/settings";
 import { ensureMetabaseProviderPropsStore } from "embedding-sdk-shared/lib/ensure-metabase-provider-props-store";
 import { mockIsEmbeddingSdk } from "metabase/embedding-sdk/mocks/config-mock";
 import {
@@ -6,6 +8,7 @@ import {
   parseDataUri,
   shouldOpenInBlankWindow,
 } from "metabase/lib/dom";
+import { createMockTokenFeatures } from "metabase-types/api/mocks";
 
 describe("parseDataUri", () => {
   it("parses a valid text data URI", () => {
@@ -77,8 +80,12 @@ describe("getUrlTarget", () => {
 describe("open()", () => {
   beforeEach(async () => {
     await mockIsEmbeddingSdk();
-    // Ensure a clean store before each test
     ensureMetabaseProviderPropsStore().cleanup();
+
+    mockSettings({
+      "token-features": createMockTokenFeatures({ embedding_sdk: true }),
+    });
+    setupSdkPlugins();
   });
 
   afterEach(() => {
